services/auth_service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
PyJWT==2.8.0

